Add navigation in the main menu
Now it's possible to use the keys Home, End, ArrowLeft and Arrow Right to navigate in the main menu. It's also possible to use Enter, Space, ArrowDown and ArrowUp to open submenus.

Clicking in a menu item also moves the 'aria-current' attribute to that item.

Still isn't possible to use Tab or Shift+Tab to jump outside the menu.